Move compiler wrapper into subpackage
